Split DNS servers into dedicated crates
