+++ rust/Cargo.toml
@@ -4,6 +4,8 @@ members = [
   "connlib/clients/android",
   "connlib/clients/apple",
   "connlib/clients/shared",
+  "connlib/l4-tcp-dns-server",
+  "connlib/l4-udp-dns-server",
   "connlib/model",
   "connlib/snownet",
   "connlib/tunnel",
@@ -158,6 +160,8 @@ firezone-telemetry = { path = "telemetry" }
 firezone-headless-client = { path = "headless-client" }
 firezone-gui-client-common = { path = "gui-client/src-common" }
 snownet = { path = "connlib/snownet" }
+l4-udp-dns-server = { path = "connlib/l4-udp-dns-server" }
+l4-tcp-dns-server = { path = "connlib/l4-tcp-dns-server" }
 dns-over-tcp = { path = "dns-over-tcp" }
 firezone-relay = { path = "relay" }
 connlib-model = { path = "connlib/model" }
