+++ rust/connlib/l4-tcp-dns-server/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "l4-tcp-dns-server"
+version = "0.1.0"
+edition = { workspace = true }
+license = { workspace = true }
+
+[lib]
+path = "lib.rs"
+
+[dependencies]
+anyhow = { workspace = true }
+domain = { workspace = true }
+futures = { workspace = true }
+tokio = { workspace = true }
+
+[dev-dependencies]
+tokio = { workspace = true, features = ["process"] }
+
+[lints]
+workspace = true
